Print usage and exit() on unrecognize options
Currently both getopt() and main.c will print an error message if there
is an unrecognized option. But this error is ignored and processing
continues as normal. This makes it very easy to miss unknown option
error messages, especially when lots of checks fail. So instead, when an
option is unrecognized, rely on getopt to print specific errors about
the unknown option, print the usage, and then exit.