Rewrite tests Makefile.am to be more maintainable and scalable.
Now the variables above the program specific SOURCES and LDADDs enumerate the files from src that are included, and the ones in the program specific SOURCES and LDADDs enumerate with src files are need for the ones in tests.  It's still clunky and there's almost certainly a better way to do this (autotools should handle dependency resolution for you), but I'm not sure how to get it working with libcheck.