Move _exemption back into the list of transform interface suffixes.
This had been on the list, and then I removed it in commit 60322c6fde252808d90564e02986a20366cfc930.  The rational for that change is not clear.  In general, this is a tricky one since both usages seem common in refpolicy.  Treating such interfaces as transform interfaces does seem to be the more common case in current upstream refpolicy though, so I'm going with that for now.