Assume that system_u and object_r are declared if no config file is loaded.
If a config file is loaded, then the users and roles to be assumed are set via "assume_users" and "assume_roles".  If no config file is loaded, then selint doesn't know about these and reports issues E-003 and E-004 on every fc file line that mentions them.  Assuming normal names is a saner behavior in the lack of a config file that should help users get more helpful results