Don't include interface calls to templates in template declarations
When an interface calls a template, this is invalid interface usage, and
generates an S-004 result.  However, if that interface calls templates,
it will not substitute the interface arguments in the template call,
resulting in the template map thinking that types named $1_t and similar
are valid types.  This can then trigger W-002 in templates.

Resolving the S-004 result also removes the W-002 warnings, but the
false positives can be confusing for users.

By not inserting interfaces calls in the template declarations, we avoid
treating types declared in interfaces as valid, which silences these
W-002 results.  These declarations were weird anyways, since they were
associated with the module of the interface rather than the module of
the te file calling the interface.