Fix passing of argument with incompatible pointer type to fts_open()
With GCC 14, the incompatible-pointer-types warning will turn into an
error by default. Due to this, the compilation of selint with newer
versions of GCC would fail.

This patch fixes 2 calls to fts_open() where this error occurs. As
fts_open() expects an array of `char * const`, not `const char * const`,
and the documentation does not guarantee that fts_open() will not modify
the string, the argument needs to be copied (or initialized as an array
instead of a pointer).

Signed-off-by: Juraj Marcin <juraj@jurajmarcin.com>